djangoapp/requirements.txt
Django>=4.2.6,<4.3
Pillow>=10.0.1,<10.1
django-summernote>=0.8.20.0,<0.8.21
django-colorfield>=0.10.1,<0.10.2
python-dotenv>=1.0.0,<1.1
django-axes==6.1.1,<6.2